Rename src to MysteryOnline, fix imports
