+++ MysteryOnline/dicegame.py
