+++ MysteryOnline/exceptions.py
@@ -1,0 +1,3 @@
+#custom exceptions go here
+
+
