+++ MysteryOnline/location.py
