+++ MysteryOnline/sprite_organizer.py
