+++ MysteryOnline/utils.py
