--- ooc_log.txt
@@ -1,59 +1,0 @@
-<02:15:24 09-11-2019> YourUsernameHere: asdad
-<02:16:02 09-11-2019> YourUsernameHere: ;ok boomer
-<02:16:08 09-11-2019> YourUsernameHere: ;ok boomer
-<02:16:09 09-11-2019> YourUsernameHere: ;ok boomer
-<02:16:18 09-11-2019> YourUsernameHere: ; shittard
-<02:16:18 09-11-2019> YourUsernameHere: ; shittard
-<02:16:26 09-11-2019> asscheeks: ;asscheeks
-<02:16:26 09-11-2019> asscheeks: ;asscheeks
-<02:19:44 09-11-2019> YourUsernameHere: whoooo
-<02:19:50 09-11-2019> assboomer: shit
-<02:19:50 09-11-2019> assboomer: shit
-<02:19:57 09-11-2019> assboomer: AAAAAAAAA
-<02:19:57 09-11-2019> assboomer: AAAAAAAAA
-<02:21:08 09-11-2019> YourUsernameHere: fuck
-<02:21:16 09-11-2019> asschooks: frick
-<02:21:16 09-11-2019> asschooks: frick
-<02:21:21 09-11-2019> asschooks: AAAAAAAA
-<02:21:22 09-11-2019> asschooks: AAAAAAAA
-<02:22:30 09-11-2019> YourUsernameHere: SHIT
-<02:23:01 09-11-2019> asschisk: reeee
-<02:23:01 09-11-2019> asschisk: reeee
-<02:23:13 09-11-2019> asschisk: reeeeeee
-<02:23:14 09-11-2019> asschisk: reeeeeee
-<02:23:17 09-11-2019> YourUsernameHere: erererere
-<02:23:18 09-11-2019> YourUsernameHere: erererere
-<02:24:04 09-11-2019> YourUsernameHere: if you gave me ass
-<02:24:18 09-11-2019> asdasdadasd: ass
-<02:24:28 09-11-2019> YourUsernameHere: she assed me
-<02:28:07 09-11-2019> YourUsernameHere: saf
-<02:28:09 09-11-2019> YourUsernameHere: asdfdsaf
-<02:28:18 09-11-2019> asdasdsadasda: thats's right
-<02:29:12 09-11-2019> YourUsernameHere: asdad
-<02:29:32 09-11-2019> YourUsernameHere: ass
-<02:29:32 09-11-2019> YourUsernameHere: ass
-<02:29:36 09-11-2019> YourUsernameHere: lol
-<02:30:08 09-11-2019> ZumoricaDev: lol
-<02:31:03 09-11-2019> ZumoricaDev: ass
-<02:31:14 09-11-2019> ZumoricaDev_: ass
-<02:32:55 09-11-2019> ZumoricaDev_: ass
-<02:33:01 09-11-2019> ZumoricaDev: geeee
-<02:33:07 09-11-2019> ZumoricaDev: hello
-<02:33:09 09-11-2019> ZumoricaDev: ass
-<02:33:10 09-11-2019> ZumoricaDev: ass
-<02:34:58 09-11-2019> ZumoricaDev_: AAAAAAa
-<02:35:43 09-11-2019> ZumoricaDev: gay
-<02:35:43 09-11-2019> ZumoricaDev: gay
-<02:35:49 09-11-2019> ZumoricaDev: awnt2die
-<02:35:52 09-11-2019> ZumoricaDev: over and over again
-<02:35:52 09-11-2019> ZumoricaDev: over and over again
-<02:35:54 09-11-2019> ZumoricaDev: geee
-<02:35:54 09-11-2019> ZumoricaDev: geee
-<02:35:58 09-11-2019> ZumoricaDev_: fuck
-<02:35:59 09-11-2019> ZumoricaDev_: fuck
-<02:36:00 09-11-2019> ZumoricaDev_: you
-<02:36:00 09-11-2019> ZumoricaDev_: you
-<02:36:06 09-11-2019> ZumoricaDev: <3
-<02:36:07 09-11-2019> ZumoricaDev: <3
-<15:15:30 09-11-2019> YourUsernameHere: gay
-<15:15:31 09-11-2019> YourUsernameHere: gay
