--- src/exceptions.py
@@ -1,3 +1,0 @@
-#custom exceptions go here
-
-
